requirements.txt
numpy>=1.26
scikit-learn>=1.4
pandas>=2.2
sentence-transformers>=2.7
faiss-cpu>=1.8
openai>=1.30
datasets>=2.19
